refactor api dir
