+++ pkg/kube/controllers/deployflow/controllers_suite_test.go
@@ -30,7 +30,7 @@ import (
 	logf "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
-	appsv1alpha1 "github.com/triton-io/triton/api/v1alpha1"
+	appsv1alpha1 "github.com/triton-io/triton/apis/apps/v1alpha1"
 	//+kubebuilder:scaffold:imports
 )
