+++ pkg/services/base/kubernetes.go
@@ -1,9 +1,9 @@
 package base
 
 import (
-	tritonappsv1alpha1 "github.com/triton-io/triton/api/v1alpha1"
 	"sort"
 
+	tritonappsv1alpha1 "github.com/triton-io/triton/apis/apps/v1alpha1"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	apiequality "k8s.io/apimachinery/pkg/api/equality"
